Fix occasional r122/ydb210 subtest failure due to REPLINSTFROZEN errors in source server log (#357)
We had a r122/ydb210 test failure in internal testing with the following diff

> ##FILTERED##... ... .. ..:..:.. 20.. : %YDB-E-REPLINSTFROZEN, Instance INSTANCE1 is now Frozen

The test greps for any errors in the source server log and expects none. But it found a
REPLINSTFROZEN error in this case. This type of error is possible in the source server log in test runs
with the debug build because the test framework induces such errors randomly (using random
gtm_test_fake_enospc env var setting). The test framework excludes this (amongst a few other errors)
when it greps for errors at the end of a subtest run.

Since ydb210 subtest explicits searches for errors in source server log files, it needs to explicitly
exclude this error to avoid false failures.