New r124/ydb324 subtest and modified v55000/GTM6940 subtest (tests YottaDB/YottaDB#324) (#356)
The main change is the addition of the r124/ydb324 subtest where the indirection
usage is tested in direct mode to ensure process does not get terminated.

During discussions of a previous round of #324 code fixes, a test case (XECUTE "ZMESSAGE 2")
assert failed and so that is also added to the r124/ydb324 subtest even though this worked
fine before the #324 code fixes.

With the #324 code changes, the v55000/GTM6940 subtest had different output.
While analyzing it, noticed that the expect file it used had lots of "send -s" commands
which sent all the commands very slow causing the test to run a lot slower (take minutes
instead of seconds) and I did not think any of this was pertinent to the test so removed
all of that. And then fixed the reference file to reflect the new output (from #324 code
changes). The new output correctly prints the YDB> prompt whereas it previously did not
and it prints a few extra errors (ZINTDIRECT and ERRWZINTR). They are due to the error trap
handler (default $ZTRAP="BREAK") being invoked now which is considered more in line with
the existing documentation for $ZINTERRUPT.