New v63005/gtm5059 subtest (tests GTM-5059 in V63005) (#358)
Tests the following release note:

YDB recognizes setting the environment variable gtm_mstack_crit_threshold to specify an integer between 15 and 95 defining the percentage of the stack which should be used before YDB emits a STACKCRIT warning. If the value is below the minimum or above the maximum YDB uses the minimum or maximum respectively. The default is 90. (GTM-5059)