Fix v63005/gtm5059 subtest failure (env var unset case incorrectly scripted)
The primary issue is that in the "unset" case, the env var gtm_mstack_size_threshold
needs to be unset (using "unsetenv") but a different env var gtm_mstack_size was being
unset (and incorrectly using "unset").

While at this, enhanced the script to also display the expected and actual recursion depth
values if/when a test scenario fails.