New v63005/gtm8943 subtest (Tests GTM-8943 in V63005) (#371)
Tests the following release note:

A ZGOTO 0 in case of call-in, unwinds all the M stack frames and returns to the invoking C routine. In case of a non-CI invocation, this terminates the process. A gtm_ci invocation without calling gtm_init() first, proceeds with the CI invocation. Previously, this produced a segmentation violation (SIG-11) error. This issue was only observed in the GT.M development environment, and was never reported by a user. (GTM-8943)