Cosmetic changes to instream.csh in v63002/v63003/v63004/v63005 subtests
The comment lines describing the purpose of each subtest are reordered now to be in the
same order as how the subtests are ordered in subtest_list_non_replic/subtest_list_replic.

Also used subtest_list_common wherever appropriate (if a subtest run -replic and not).

And reinserted a comment line about gtm8980 that was lost in a prior commit.

Fixed outref.txt wherever necessary.