Fix v63005/gtm5059 test failure in pro build
The v63005/gtm5059 test passed with debug builds but failed with production builds as follows.

Recursion depth does NOT match expected. Expected depth = 87268.2280701754385 : Actual depth = 87267

The test allows for a very tight margin of error in the recursion depth (only 1 max between actual
and expected depth). But the actual difference was 1.22 (slightly over 1). When the actual recursion
depth is of the order of 87000, a difference of 1 or 2 should not matter. So the test is now fixed
to allow for an error margin of 2 on the lower side (and continue to maintain an error margin of 1
on the upper side) this way both pro and dbg builds run this test fine.