Fix relink/keeprtn subtest failure due to issue in gen^randbool
We had one failure in the relink/keeprtn subtest (happened once in 100 runs across various hosts)
where gtm_autorelink_keeprtn env var was set to FALSE but the dummy.o object stayed in the
rtnobj shared memory which was not expected.

Turns out the test sets the env var to FALSE using a call to 'mumps -run gen^randbool 0'. This
returns one amongst a lot of choices each of which should evaluate to the equivalent of FALSE.
But in some cases, it could return strings that start with a 0 (to denote FALSE) but followed by
a string of decimal digits that evaluates to a non-zero number (e.g. 018). This caused the
keeprtn subtest to treat the env var setting of 018 as if a TRUE setting was specified and fail.

The fix is to randbool to check if the string it is about to return (in the type=0 case)
evaluates to a TRUE and if so keep retrying with another random string instead. Just like is
already done in the if (""=retx) part of the code.