[!347] Fix expect/terminal subtest failure now that $Y is maintained correctly in case of wrap
When terminal deviceparameter LENGTH is set to 2 and WIDTH is set to 30, the output of a long
line used to show up as follows.

what is LENGTH in effect acros
s one write statement that spa
ns multiple lines: 2some more2
even more3

The two occurrences of 2 and one occurrence of 3 in the last two lines of the above output
are $Y printed as part of the WRITE command. Since LENGTH is set to 2, $Y can never be >= 2.
The above output where $Y is 3 is proof that $Y could even go more than LENGTH before.

With the !347 code fixes, $Y stays below 2 in the above so the revised output reads as

what is LENGTH in effect acros
s one write statement that spa
ns multiple lines: 0some more0
even more1

The 2 is replaced by 0 and 3 by 1. This is the correct output where $Y is always < 2.
The reference file is fixed to reflect the new output.