Fix rare v54001/C9J04003116 subtest failure
We had an in-house failure of the v54001/C9J04003116 subtest with the following diff.

22a23
> gzip: mumps.mjl: file size changed while zipping

The test pauses updates for a short duration and in that time frame zips and unzips the journal file.
It is during the zip that we got the above error.

Analysis showed that the process which paused the updates (pid = 25470 below) did a flush of the
db/jnl buffers by invoking a DSE process (pid = 25494 below) to do DSE BUFF. Below is a detailed
journal extract output of relevant portions of the mumps.mjf journal file.

   1880 0x00034b98 [0x00d0] :: PINI   \64879,57122\17179870085\144567650\25470\thyme\nars\\0\\\
   1881 0x00034c68 [0x0048] :: SET    \64879,57122\17179870085\3225565931\25470\0\0\0\0\0\0\^%pauseflag="1"
   1882 0x00034cb0 [0x0058] :: SET    \64879,57122\17179870086\2964196033\25411\0\0\0\0\0\0\^%pauseflag(0,"pausedcount")="1"
   1883 0x00034d08 [0x0058] :: SET    \64879,57122\17179870087\3675728647\25413\0\0\0\0\0\0\^%pauseflag(0,"pausedcount")="2"
   1884 0x00034d60 [0x0058] :: SET    \64879,57122\17179870088\2890707879\25405\0\0\0\0\0\0\^%pauseflag(0,"pausedcount")="3"
   1885 0x00034db8 [0x0058] :: SET    \64879,57122\17179870089\2310738196\25407\0\0\0\0\0\0\^%pauseflag(0,"pausedcount")="4"
   1886 0x00034e10 [0x0058] :: SET    \64879,57122\17179870090\1315868676\25409\0\0\0\0\0\0\^%pauseflag(0,"pausedcount")="5"
   1887 0x00034e68 [0x00d0] :: PINI   \64879,57123\17179870091\2318067546\25494\thyme\nars\\0\\\
   1888 0x00034f38 [0x00b8] :: EPOCH  \64879,57123\17179870091\675521579\25494\0\0\0\972\1027\1
   1889 0x00034ff0 [0x0020] :: PFIN   \64879,57123\17179870091\3264249108\25494\0
   1890 0x00035010 [0x0020] :: PFIN   \64879,57123\17179870091\2251730520\25470\0

The DSE BUFF would only flush the journal file until the EPOCH record. The PFIN record written
by pid=25494 and pid=25470 are not guaranteed flushed at the time "do pause^pauseimptp" returns
back to C9J04003116.csh. This means that it is possible one of the paused imptp jobs could flush
these PFIN records to the journal file a little later. And since the PFIN record of pid=25494
happens to cross over a filesystem boundary of 0x1000 (starts at 0x34ff0 and ends at 0x35010)
writing that to the journal file on disk would cause the journal file size to change and if the
test script C9J04003116.csh is doing a gzip at the time this flush happens we would see the above
error from gzip.

The fix is to make sure C9J04003116.csh does a VIEW "JNLFLUSH" after "do pause^pauseimptp" returns.
This way the PFIN records are flushed too and we are guaranteed a stable journal file before we
attempt the gzip.