Fix occasional r124/ydb293 subtest failure
The failure symptom was that the ZWRITE of ^a and ^b that the test did on INST2 showed up nothing.
The test waits a hardcoded 1 second for the updates on INST1 to get replicated across to INST2.
This is not guaranteed to work right on slower systems that can sometimes even take a minutes to
replicated all the data.

This is now fixed by doing the cleaner MSR SYNC which would return only after all updates from
INST1 are replicated to INST2.

While at this, pre-existing dbcreate and dbcheck invocations are checked for error status.