Fix rare expect/tterase subtest failure
We had one failure of this subtest with the following failure diff

1c1
< 56 terminal interactions involving ERASE special input character, BACKSPACE and DELETE keys are successsful.
---
> 55 terminal interactions involving ERASE special input character, BACKSPACE and DELETE keys are successsful.

Which interaction was missing was found by looking at tterase.exp.output.
Below is the normal content of this file in case of a pass test run.

tterase.exp.output
    236 case:(NOESCAPE:NOEDIT:NOEMPTERM)
    237
    238 Enter ERASE special terminal input character on nonempty read buffer - Use string GTMM - followed by newline character
    239 GTMM^H ^H
    240 PASS
    241 Enter BACKSPACE key on nonempty read buffer - Use string GTMM - followed by newline character
    242 GTMM^H ^H
    243 PASS
    244 -------------------------------------------------------------------------
    245 GTM7675>exit^M
    246 exit

In the failure run, lines 243-246 were missing. This caused one PASS to miss in the total count
resulting in only 55 PASS occurrences instead of the expected 56 occurrences.

The timeout was set at 10 seconds in this test and no timeout handler was set up so I suspect in the failure
run the host was very slow and it took more than 10 seconds to return back to the prompt causing an expect
script timeout which did not show up in the final output because of the way the test scripts were structured.

The fixes done were many fold.

1) The timeout is now bumped to 60 seconds. This is the primary fix.
2) The expect script structure is revised to reflect recent expect coding standards.
   This should make the reference file a lot more readable than before as well as let us better debug failures if any.