Fix rare spanning_nodes/basic subtest failure
We had one failure in-house where in the "#Start zshow" section of the reference file, the zshow "G"
output had DRD of 4 whereas the reference file was expecting a value of 3.

spanning_nodes/inref/zshow.m is the M program that is invoked for this section. It does a set of
99 updates each setting a node of size 2500 bytes. It is not exactly clear how DRD (the # of blocks
read in from disk) can be different than 3 (which is because we need to read blocks 0, 1, 2 all once).
This means some block was read more than once. I suspect it is the bitmap block. But not sure why.
Maybe slowness of the armv6l box caused this. Nevertheless, the primary purpose of the test is to
verify most of the zshow statistics. The test already masks 3 gvstats which can have varying output.
So DRD is now added to that list to avoid failures like this.