Fix possibility of hang/spin-loop in v63003/gtm8735 subtest
In one test run, the v63003/gtm8735 subtest was found to be hung.
Turns out it was in a while loop (which was a spin-loop) attempting to change
the help database gtmhelp.dat to NOREAD_ONLY and encountering READONLYLKFAIL errors.
And that was because another concurrently running v63002/gtm8165 subtest had this
help database file open. That was hung for a different reason and so it caused the
v63003/gtm8735 subtest to also hang/spin-loop in the while loop.

This is now fixed so we do not hang eternally but wait for a max of 300 iterations
and sleep loop instead of spin loop by sleeping 1 second in between iterations.

And if the test detects a 300 second time out did occur, it notes down the pids
that have the help database file open and displays them too for further debugging.