Fix test failures due to TEST-E-FAIL occurrence from concurrent tests in ps -ef listing
The ideminter_rolrec/mupipstop_rollback_or_recover subtest failed with the following diff

> ideminter_rolrec_0_1/mupipstop_rollback_or_recover/MUPIP_RECOVER.log
> user 8788 8587 0 17:10 pts/1 00:00:00 mumps -run %XCMD for i=1:1:120 write:i=120 "TEST-E-FAIL No update by the bg process",! quit:$data(^GBL)!(i=120) hang 0.5
> ideminter_rolrec_0_1/mupipstop_rollback_or_recover/mstop_17_10_04.log
> nars 8788 8587 0 17:10 pts/1 00:00:00 mumps -run %XCMD for i=1:1:120 write:i=120 "TEST-E-FAIL No update by the bg process",! quit:$data(^GBL)!(i=120) hang 0.5

The pid identified was a concurrently running v62000/gtm7976 subtest which had a TEST-E-FAIL
usage in its ps -ef listing. The test framework checks for -E- usages in *.log files and so
that caused the test failure.

The fix to prevent this is to move mumps -run %XCMD usages of existing tests which have the
use of TEST-E-FAIL to inside an M program that way ps -ef listing is not affected with -E-
occurrences. There were 2 such subtests and both are now fixed.