Fix v60000/gtm6571 subtest failure on slow systems due to gdb attach/detach taking 2 to 3 minutes
We had a test failure on an ARMV6L box with the following diff.

> TEST-E-FAIL, MUTEXLCKALERT message appeared 5 times in syslog.

The test is structured so a MUTEXLCKALERT message is seen in the syslog every 35 seconds or so.
And it expects to see 3 such messages in the syslog for the duration that it runs. But this test
invokes the gtm_procstuckexec scheme which will attach to the crit holding process using gdb
and get its C-stack. In the test failure, this process took 2 minutes and 35 seconds (from
02:49:09 to 02:51:43). Normally this is expected to take less than a second.

	TRACE_MUTEXLCKALERT_7055_7046.outx.gz
	--------------------------------------
	BLOCKING PID = 7046
	WAITING PID = 7055
	Time before calling dbx: Tue Aug 28 02:49:09 EDT 2018
	Count is: 3
	.
	.
	Now the time is: Tue Aug 28 02:51:43 EDT 2018 dbx exit status: 0

In the same time, the test generated 5 MUTEXLCKALERT messages which was more than the max of 4 it allowed.
And as a result the test failed.

We have previously seen gdb take a long time on the 1-CPU systems. So this subtest is disabled
there. Also, the test has now been made stricter by searching for exactly 3 MUTEXLCKALERT messages.
Previously it used to allow anywhere from 1 to 4.