Fix occasional r122/tprestart test failure by making it tn-based instead of time-based (!394)
We had a failure of this subtest on a slow box with the following diff.

> No TPRESTART messages in the syslog from this test run. The test is not serving its purpose. Signal failure.

In this case, in the 10 seconds that the subtest waited for 5 mumps processes to do TP updates,
only 5 TSTART/TCOMMIT updates happened. This was not enough to create the contention needed to
get any TPRESTART messages in the syslog.

The test has now been reworked to be tn-based. That is, each child waits for at least 100 TP
transactions to be done before being ready to terminate. Once all 5 children are done 100 TP
transactions, all children will terminate. This should be enough to reliably produce TPRESTART
messages that this test cares about. On faster systems, this test will take less than 10 seconds
and on slower systems this test might sometimes take more than 10 seconds.

Also ran the revised test with r120 (the version which did not have the particular code fix this
test is testing) and verified it saw TPRESTART messages where global names did not match the
mapped db file name that was reported. So the revised test does serve its purpose too.