[#350] Fix occasional v55000/GTM6940 test failure (more frequent after https://gitlab.com/YottaDB/DB/YDB/issues/350)
The v55000/GTM6940 test failed with an extra new line in the diff more than 50% of the time
in test runs after the #350 code changes.

The test had a pre-existing timing issue in that it was not waiting for the YDB> prompt after
waiting for the ERRWZINTR message to show up. For reasons not yet clear, this was not an issue
until the #350 changes to do tcsetattr() calls differently. Most likely that changed the timing
of terminal output enough to increase the likelihood.

The timing issue is now fixed in the .exp script by waiting for the YDB> prompt and the test
passes fine after that.

While changing the .exp file, the .csh file was also enhanced to do error checking from the
"expect" command invocation.