[#350] Fix occasional v62001/gtm8206 subtest failure (more frequent after https://gitlab.com/YottaDB/DB/YDB/issues/350) (!400)
The v62001/gtm8206 subtest failed occasionally with the following diff on slower boxes
particularly when multiple tests were running concurrently.

2c2
< %YDB-I-CTRLC, CTRL_C encountered
---
> %YDB-I-CTRLC, CTRL_C encounteredhalt

The issue is that the expect script was waiting for CTRL_C to show up.
And then issuing the "halt" message. But it was possible for the "halt" to be sent
even before the newline was printed by the CTRL_C printing logic (which would soon also
print the YDB> prompt). So the fix is to wait for the YDB> prompt instead of wait for
CTRL_C error to show up. This avoids all timing issues.

While at that, some more cleanup of the expect script was done (to avoid code specific to
non-supported platforms) and add expect debugging.