Fix occasional v52000/D9H02002642 subtest failure (more frequent after https://gitlab.com/YottaDB/DB/YDB/issues/350) (!402)
The v52000/D9H902002642 subtest failed with the following diff once in a while on slow systems.

4c4
< x=27
---
> ^[[23~x=27

This subtest enters the following set of commands at the YDB> prompt.

YDB>use $p:escape  r *x s z=$zb  zwrite

Due to recent #350 code changes, the terminal has ECHO setting turned on except for the
duration of the "r *x". So if the escape sequence that the test sends (equivalent to F11)
happens after the USE but before the READ * is invoked, the escape sequence will be echoed
on the terminal as is now whereas it did not get echoed before the #350 code changes.

Depending on when the F11 keypress happens relative to the execution of the USE and READ *
commands, we could see the extra "^[[23~" output in the x=27 line which explains the above diff.

To avoid this failure, the test has been reworked so it is in line with various other expect
script usage in the test system. And a "write !" was introduced before the ZWRITE so the
escape sequence, if at all it shows up, is displayed in a separate line than the x=27 line.

But after the rework, the occasional test failure too disappeared. Not sure exactly why. But
if it happens again, all that is needed is to filter out the line containing just the
escape sequence from expect_sanitizied.out in D9H02002642.csh. Will handle that if/when the
next such failure occurs.