[#350] Fix occasional expect/terminal subtest failure (more frequent after https://gitlab.com/YottaDB/DB/YDB/issues/350)
We had failures of this test in slow boxes more than 50% of the time. This is because the terminal
has ECHO setting ON for the most part of the mumps process execution in direct mode which means
a newline sent by the expect script would echo on the terminal now whereas it did not before the #350
code changes thus resulting in a test failure due to differing line #s where the desired pattern
string is found.