Fix occasional v63003/gtm8889 subtest failure
This test failed occasionally with the following diff on heavily loaded boxes

4a5
> EXPECT Failed
23,25c24
< Topic?
< YDB>halt
< SHELL
---
> Topic? timeout occurred

In this portion of the test, we wait for "Topic?" to show up, then send a Ctrl-C
followed by a \r and another \r all in one sequence and wait for a "YDB>" prompt
to show up. Instead of the "YDB>" prompt, in the failure, we see a "Topic?" prompt.

I suspect this is because all these keystrokes are entered when the ZHELP is not yet
gone into the READ command to wait for input. That causes some keystrokes to be
swallowed causing different output.

This is now fixed by moving the wait_for_stty_echo_off.csh script from the r124 test
to "com" directory. This is a generic enough script that it will be used by various
subtests so belongs in the test framework.

In addition, com/expectinit.exp, a central script that every expect test is expected
to source, is enhanced to note down the terminal name in a file term_env.txt which
will later be used by wait_for_stty_echo_off.csh if later invoked in the expect script.

The gtm8889.exp script is then enhanced to call wait_for_stty_echo_off.csh before sending
the Ctrl-C. This avoids all timing issues.

Additionally, because of the changes to com/expectinit.exp, a few other tests that use
this script have an extra line in their reference files.