Fix occasional v63004/gtm5730 subtest failure due to slow test runtimes
We had a test failure once in an in-house box in a debug test run where a core file got created but
otherwise the test passed fine. The reason for the core file was that an fsync() took too long
resulting in a timer pop signaling a timer was long overdue (at gt_timers.c line 732).

 #0  0x00007f7c31ed7187 in kill () at ../sysdeps/unix/syscall-template.S:78
 #1  0x00007f7c311756e1 in gtm_dump_core () at sr_unix/gtm_dump_core.c:69
 #2  0x00007f7c31176bd6 in gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:148
 #3  0x00007f7c3117132f in timer_handler (why=14) at sr_unix/gt_timers.c:732
 #4  <signal handler called>
 #5  0x00007f7c31faf214 in fsync (fd=5) at ../sysdeps/unix/sysv/linux/fsync.c:27
 #6  0x00007f7c316a99ce in jnl_fsync (reg=0x55d6c5906170, fsync_addr=189856) at sr_unix/jnl_fsync.c:134
 #7  0x00007f7c313b6ea2 in jnl_wait (reg=0x55d6c5906170) at sr_port/jnl_wait.c:39
 #8  0x00007f7c314816e0 in op_tcommit () at sr_port/op_tcommit.c:552

The reason the fsync() happened is because TSTART by default does hardening of journal file.
Specifying BATCH in the TSTART avoids this and since that speeds up the test without affecting
the purpose of the test, that is the change done to fix such failure.