Fix occasional r124/ydb350 subtest failure (more frequent after https://gitlab.com/YottaDB/DB/YDB/issues/350)
We had two failures till now in this subtest. In both outputs below, the 1 is not expected but shows up.
In the first output it shows up in a new line. In the second output it shows up in the same line.

Output1
-------
strace -o trace2.outx $ydb_dist/mumps -run readcmd^ydb350
1
; Below are terminal settings BEFORE READ command

Output2
-------
strace -o trace3.outx $ydb_dist/mumps -run readstarcmd^ydb350
1; Below are terminal settings BEFORE READ * command

The reason for this is that we send the "1" as an input to readcmd^ydb350 and readstarcmd^ydb350
after issuing a "mumps -run readcmd^ydb350" etc. But it is possible when we do the send of "1",
the mumps invocation has still not reached the READ or READ * command which means the terminal
ECHO setting would still be turned ON. That results in the "1" being displayed.

To avoid this issue, the test is now fixed to wait until the terminal ECHO setting is turned OFF
before sending the "1" as input.