Fix occasional v63002/gtm8694 subtest failure due to user input getting incorrectly displayed (related to https://gitlab.com/YottaDB/DB/YDB/issues/350)
We had the v63002/gtm8694 subtest fail with the following diff.

97c97
< ##DIFFERENT EXPRESSION##^.*YDB-I-CTRLC, CTRL_C encountered$## VS ##%YDB-I-CTRLC, CTRL_C encounteredwrite $char(65,66,67,68),!##
---
> %YDB-I-CTRLC, CTRL_C encounteredwrite $char(65,66,67,68),!
141c141
< ##DIFFERENT EXPRESSION##^.*YDB-I-CTRLC, CTRL_C encountered$## VS ##%YDB-I-CTRLC, CTRL_C encounteredwrite $char(65,66,67,68),!##
---
> %YDB-I-CTRLC, CTRL_C encounteredwrite $char(65,66,67,68),!

The issue here is that after the CTRLC message shows up, the user input of "write $char(65,66,67,68),!"
shows up in the same line whereas we expect that to not get displayed because it should be entered when
we are in direct mode read. Until we go into direct mode read, terminal ECHO is turned ON. So we need
to wait for terminal ECHO to be turned OFF before sending the user input.

This is done by the wait_for_stty_echo_off.csh script invocation.