Fix occasional r120/ctrlchandler subtest failure (related to https://gitlab.com/YottaDB/DB/YDB/issues/350)
The r120/ctrlchandler subtest failed a few times with the following diff

24c24,25
< YDB>set savelvl=$zlevel set $etrap="do etraperror^ctrlchandler"
---
> YDB>set savelvl=$zlevel set $etrap="do etraperror^ctrlchandler"use $Principal:(CTRAP=$char(3):exception="do gotctrlc^ctrlchandler":CENABLE)
>

Two lines of user input showed up in the same line because the second line of input was sent
through the expect script before the YDB> prompt showed up after the first line was processed.
This timing issue is fixed by waiting for the terminal ECHO setting to be turned OFF before
sending each line of user input.