Fix occasional simpleapi/tp subtest failure due to TPTIMEOUT taking 2 seconds instead of 1 second
We saw this test fail in one slow box where the timeout took 2 seconds instead of 1 second.
The C program simpleapi/inref/tp5_TPTIMEOUT.c already allows 1 or 2 seconds and treats them
as a PASS and only treats 3 seconds or more of elapsed time as a FAIL. But the reference file
only handles elapsed time of 1 second. So this is now fixed to allow 1 or 2 seconds in the PASS case.