Fix v52000/D9G12002636 subtest failure due to $SHELL not hardcoded to /usr/local/bin/tcsh
Due to a recent rework of the v52000/D9G12002636 test, the entire expect script output is now
part of the reference file. The expect script had always done a spawn of $SHELL. This turns
out to be /usr/bin/tcsh on some platforms (Arch Linux in our setup). So instead this is
hardcoded to spawn /usr/local/bin/tcsh, the standard format for all tests in the test system.