Fix rare r120/readcmdrecallhist subtest failure due to timing issue in expect script
We had a test failure with the following diff.

50d49
< PASS from readcmd2_complex

This is because in readcmd2_complex.exp, we do a send at line 110 and line 112 below both
of which will produce a YDB> prompt after finishing executing but we do an expect of the
prompt only once at line 113. So it is possible the YDB> prompt we see at line 113 is
actually the prompt after executing line 110. And because of this, the flow of the test
is disrupted enough that it is possible we end up prematurely exiting the expect script
even though line 112 has not yet been executed (which is the case in the test failure).

  110         send -- "if x($i)'=\"${newstr}\" write \"Expected x($i)=$newstr but actual=\",x($i),! if \$incr(^failcnt)\r"
  111         # Increment ^passcnt if test passed
  112         send -- "if x($i)=\"${newstr}\" if \$incr(^passcnt)\r"
  113         expect -exact "YDB>"
    .
    .
  123 puts "\n# Halt"
  124 send -- "halt\r"
  125 expect -exact ">"
  126 exit

Introducing an expect of the YDB> prompt before line 112 should fix this timing issue.