Fix rare v54000/D9I09002699 subtest failure due to YDB-E-JNLCNTRL message in syslog from a concurrent test
We had a v54000/D9I09002699 test failure with the following diff.

26c26,27
< SUBTEST PASS
---
> SUBTEST FAIL
> Verify the error in Test.logx

This is because the test did find a JNLCNTRL message in the syslog whereas it does not expect to
find any. This message turned out to be from a concurrently running test (v53002/D9I07002690).

The fix to avoid such messages from causing false failures is to search for the current directory
(`pwd`) also while trying to grep for the JNLCNTRL message.