Introduce com/sstepgbl.m to single step M program and record activity in ^%sstepgbl; Enhance io/diskfollow_timeout subtest to use sstepgbl.m so we have a record of actual timing of events across multiple processes in a global in case of test failures
sstepgbl.m is modeled on sstep.m but instead of displaying each M line as it gets executed, it records
this in a global. This lets once see the timing of execution of M lines across multiple processes in
one shot. This is particularly helpful in case of the io/diskfollow_timeout subtest where there are
3 to 4 processes running concurrently (some doing READS, some doing WRITEs, some doing MUPIP INTRPTs etc.).
In case a test fails, examining ^%sstepgbl helps us understand the timing of events and better explain
why the test failed. For example, in one failure, a "hang 5" done in one of the processes took actually
15 seconds to return. This is now easily confirmed with ^%sstepgbl as it records the time too.

For now, the io/diskfollow_timeout subtest is changed to use sstepgbl.
sstepgbl.m can also be used in other tests as the need arises.