Exclude io/diskfollow_timeout subtest on hosts where env var ydb_test_exclude_diskfollow_timeout is set to 1
We noticed that on a particular ARMV7L host, the io/diskfollow_timeout subtest fails once in a while.
Although the failure diff is not the same every time, a sample diff is pasted below.

165c165
< x= ՇՈՉ length(x)= 3
---
> x=  length(x)= 0
168c168
< $device= 0  $za= 0 $test= 0 $zeof= 0 $x= 3
---
> $device= 0  $za= 0 $test= 0 $zeof= 0 $x= 0

Using ^%sstepgbl tracing, it was determined that we did a "hang 5" and that returned back to us 13 seconds
later instead of 5 seconds later. This caused the test assumptions to be broken because the test has
multiple processes one doing a 15 second timeout READ while another writer is doing two writes 5 seconds
apart. We expect the reader to see both the writes in one read because the read is a 15 second read. But
that does not happen because the 5 second sleep in between the writes turns out to be 13 seconds apart in
reality.

It is not easy to rework this test to increase the read timeout from 15 seconds to say 60 seconds (the
total test run time will change across all platforms, it might take a lot of time to figure out the
smallest timeout that will reliably pass on the failing host etc.). Therefore for now, this test is
disabled only on those specific affected hosts by setting the ydb_test_exclude_diskfollow_timeout env var
to 1 only on that host.