Fix rare v63000a/gtm4759 subtest failure (pid takes more than 10 seconds to terminate)
In one in-house test, we noticed 10 seconds was not enough for the process to die (when the
system was running 3 tests concurrently). We do not know exactly how long later that process
terminated (needs process accounting enabled etc.) but expect 5 minutes should be more than
enough for this purpose even on a loaded system.