Revise r124/ydb359 subtest to test user-level workaround now that #359 code fixes have been reverted
The user-level workaround for #359 is to ensure ZSTEP action is done inside $ZINTERRUPT code. That
will ensure the ZSTEP action is not lost inspite of MUPIP INTRPTs. The r124/ydb359 subtest, instead
of being deleted altogether, is now revised to test this workaround even though the #359 code
change has been reverted.