Fix rare mupjnl/test_verbose subtest failure due to different EOF record offset
Twice in the past year of testing, we have seen the mupjnl/test_verbose subtest fail
with the following diff.

106c106
< %YDB-I-MUINFOUINT4,     Record Offset : 75024 [0x00012510]
---
> %YDB-I-MUINFOUINT4,     Record Offset : 76472 [0x00012AB8]

The record offset displayed above is the EOF record offset.
And that turns out to be more than expected because the journal file has 3 EPOCH
records in the failure case instead of the usual 2. It is possible on a heavily loaded
system that an extra EPOCH gets written because the epoch interval of 30 seconds (in debug
test runs) is reached before the end of the test. If so that can cause extra PBLK records
to be written which can in turn increase the final EOF record offset.

In both failures that we have seen till now, the offset in the failure was the same
(0x12AB8) so the reference file is now fixed to allow for this additional offset.