[#357] Fix v63004/gtm8909 subtest reference file due to https://gitlab.com/YottaDB/DB/YDB/issues/357 code changes
Ctrl-C (i.e. SIGINT) handling is now deferred (instead of lost) while inside the HELP tool
in YottaDB due to #357 code changes. In the case of LKE, this means it prints a YDB-I-CTRLC
message indicating a Ctrl-C was pressed once control comes back to LKE from the HELP tool.

The original purpose of the v63004/gtm8909 subtest (i.e. Ctrl-C in the HELP tool does not lead
to a ENO256 error upon exit) is still not broken by this change. So the fix done here is to
the reference file to reflect new code behavior.