Fix rare v54000/C9D08002390 subtest failure due to overlap of time windows of different sections in the test
We had a rare failure in the v54000/C9D08002390 subtest with the following diff.

40a41
> YDB-I-STUCKACT, Process stuck script invoked: FAILURE : v54000_0/C9D08002390/gtmprocstuck_get_stack_trace.csh WRITERSTUCK 7949 7949 1 -- generated from 0x00007F0CBA88B245.
41a43
> YDB-E-WRITERSTUCK, Buffer flush stuck waiting for [0x00000001] concurrent writers to finish writing to database file v54000_0/C9D08002390/mumps.dat

There was an extra STUCKACT and WRITERSTUCK message seen in the Test case 3 section of the test.
This was because the "Time after test case 2" (recorded in the test output) was identical to the
"Time before test case 3". And so the getoper.csh invocation in Test case 3 also got syslog messages
that were generated in Test case 2.

To avoid such failures, a "sleep 1" is done at the start of Test case 2 and 3.

While at this, it was noticed that the "after" times were never being used so they are now removed.