Disable timing/gtm8680 subtest on ARM boxes as it needs lots of processes which makes response time non-deterministic
When the gtm8680 subtest was in the v63003 test (which is part of E_ALL), it was disabled on the
ARM boxes.  But when it got moved to the timing test (which is part of M_ALL), it was re-enabled
in the hope it might work fine. But it did not so disabling it back in the ARM boxes now.