Fix manually_start/maxtrignames subtest to be true to its purpose
This subtest was using white-box test cases to reduce dbg test runs to just 1000 names
instead of the originally needed 1 million names. But in GT.M V6.3-005, this white-box
test case # (116) was replaced with something else (a 2-second sleep) which caused every
trigger load in this test to run horribly slow. Therefore, this test is now fixed to
use 1 million names (and in turn take a lot longer than it would for just 1,000 names)
even in dbg i.e. the white-box case is removed. So pro & dbg run the same test now.